+++ _retail_/Interface/AddOns/BigWigs_BurningCrusade/CHANGELOG.md
@@ -1,6 +1,8 @@
 # BigWigs [Burning Crusade]
 
-## [v9.2.0](https://github.com/BigWigsMods/BigWigs_BurningCrusade/tree/v9.2.0) (2022-02-23)
-[Full Changelog](https://github.com/BigWigsMods/BigWigs_BurningCrusade/compare/v9.1.1...v9.2.0) [Previous Releases](https://github.com/BigWigsMods/BigWigs_BurningCrusade/releases)
+## [v9.2.1](https://github.com/BigWigsMods/BigWigs_BurningCrusade/tree/v9.2.1) (2022-05-02)
+[Full Changelog](https://github.com/BigWigsMods/BigWigs_BurningCrusade/compare/v9.2.0...v9.2.1) [Previous Releases](https://github.com/BigWigsMods/BigWigs_BurningCrusade/releases)
 
-- Bump toc  
+- Hyjal/Trash: Tweak  
+- Hyjal/Trash: Update use of :GetGossipOptions  
+- workflow: move to checkout@v2